test: refactored tests
