+++ tests/legacy/test_hex_conversion.py
