+++ tests/legacy/test_spike_checks.py
