+++ tests/test_processing_modules.py
@@ -1,0 +1,188 @@
+import logging
+
+import numpy as np
+import xarray as xr
+import pytest
+from conftest import (
+    assert_different_np_array,
+    cnv_path,
+    test_cnv,
+)
+
+from ctdam.exceptions import BinnedDataError
+from ctdam.parser.read_ctd_data import read_ctd_data
+from ctdam.proc.modules import (
+    AirPressureCorrection,
+    AlignCTD,
+    BinAvg,
+    LoopRemoval,
+    WFilter,
+    wildedit_geomar,
+)
+from ctdam.proc.modules.seabird_functions import CellTM
+
+logger = logging.getLogger(__name__)
+
+
+@pytest.fixture
+def ds():
+    return read_ctd_data(test_cnv)
+
+
+def test_alignctd(ds):
+    oxy = ds.oxygen.copy(deep=True)
+    ds.proc.module("alignctd")
+    assert_different_np_array(oxy, ds.oxygen.values, ds)
+
+
+@pytest.mark.xfail(reason="Debug failing")
+def test_wildedit_geomar_logic(ds):
+    salinity = ds.salinity.copy(deep=True)
+    new_data, _ = wildedit_geomar(
+        data=salinity,
+        flag=ds.flag,
+        std1=3.0,
+        std2=10,
+        window_size=50,
+    )
+    assert_different_np_array(salinity, new_data, ds)
+
+
+def test_air_pressure_correction(ds):
+    old_data = ds.pressure.copy(deep=True)
+    new_ds = AirPressureCorrection()(ds=ds)
+    pressure_diff = new_ds.meta.provenance()["airpressure"][
+        "pressure_diff"
+    ].split()[0]
+    assert float(old_data[0]) + float(pressure_diff) == float(
+        new_ds.pressure[0]
+    )
+
+
+def test_airpressure_with_bugged_metadata(ds):
+    ds.meta.custom()["Air_Pressure"] = "definitely_not_a_number"
+    new_ds = AirPressureCorrection()(ds=ds)
+    assert isinstance(new_ds, xr.Dataset)
+
+
+@pytest.mark.xfail(reason="binning detection not implemented yet")
+def test_binned_data_error():
+    ds = read_ctd_data(cnv_path / "MSM140_1.cnv")
+    with pytest.raises(BinnedDataError):
+        AlignCTD()(ds=ds)
+
+
+# @pytest.mark.xfail(reason="binning detection not implemented yet")
+def test_bin_avg(ds, create_files):
+    bin_variable = "pressure"
+    if bin_variable not in ds:
+        pytest.skip()
+    new_ds = BinAvg()(
+        ds=ds,
+        arguments={
+            "bin_variable": bin_variable,
+            # "bin_size": 0.1,
+            "bin_size": 1,
+        },
+    )
+    if create_files:
+        new_ds.export.to_cnv(f"binavg_{new_ds.attrs['path_to_source_file']}")
+    diff = np.diff(new_ds[f"{bin_variable}_bins"].data)
+    # assert len(diff[np.isclose(diff, 0.1)]) > len(diff) * 0.95
+    assert len(diff[np.isclose(diff, 1)]) > len(diff) * 0.95
+
+
+def test_binavg_linear_interpolation():
+    ds = read_ctd_data(cnv_path / "EMB295_14-1.cnv")
+    sparse = BinAvg()(ds)
+    dense = BinAvg()(ds, arguments={"linear_interpolation": True})
+
+    assert len(dense.pressure_bins) >= len(sparse.pressure_bins)
+    gaps = np.diff(dense.pressure_bins)
+    assert np.allclose(gaps, gaps[0])
+
+
+def test_wfilter(ds, create_files: bool):
+    pre_ds = ds.copy(deep=True)
+    new_ds = WFilter()(
+        ds=ds,
+        arguments={
+            "tUrB iDIty": {"window_type": "median", "window_width": 10},
+            "Temperature": {"window_width": 50},
+            "Oxygen": {"half_width": 2, "offset": 5},
+            "Conductivity": {
+                "window_type": "triangle",
+                "window_width": 200,
+                "half_width": 3,
+                "offset": 3,
+            },
+        },
+    )
+    # check for boundary effects
+    assert new_ds.pressure[0] > 0.2
+    assert ds.access.size() == new_ds.access.size()
+    for param in pre_ds:
+        if param not in [
+            "pressure",
+            "temperature",
+            "conductivity",
+            "salinity",
+            "oxygen",
+            "fluorometer",
+            "turbidity meter",
+            "par",
+            "spar",
+        ]:
+            continue
+        if create_files:
+            new_ds.export.to_cnv(
+                f"wfilter_{new_ds.attrs['path_to_source_file']}"
+            )
+        assert_different_np_array(
+            pre_ds[param].data, new_ds[param].data, new_ds
+        )
+
+
+def test_cell_tm(ds, create_files):
+    if not "conductivity" in ds:
+        pytest.skip()
+    pre_ds = ds.copy(deep=True)
+    new_ds = CellTM()(
+        ds=ds,
+        arguments={},
+    )
+    if create_files:
+        new_ds.export.to_cnv(f"celltm_{new_ds.attrs['path_to_source_file']}")
+    assert_different_np_array(
+        pre_ds.conductivity.data,
+        new_ds.conductivity.data,
+        new_ds,
+    )
+
+
+def test_new_loop_flags(ds, create_files):
+    pre_ds = ds.copy(deep=True)
+    new_ds = LoopRemoval()(
+        ds=ds,
+        arguments={},
+    )
+    if create_files:
+        new_ds.export.to_cnv(
+            f"loopremoval_{new_ds.attrs['path_to_source_file']}"
+        )
+    assert_different_np_array(pre_ds.flag.data, new_ds.flag.data, new_ds)
+
+
+def test_time_dependent_loop_removal():
+
+    module = LoopRemoval()
+    # strictly increasing — nothing should be flagged
+    pressure = np.array([0.0, 1.0, 2.0, 3.0, 4.0, 4.01])
+    flags = module.time_dependent_loop_removal(pressure=pressure, delta=0.0)
+    assert not flags.any()
+
+    # loop at index 3: should be flagged
+    pressure = np.array([0.0, 1.0, 3.0, 2.0, 4.0])
+    flags = module.time_dependent_loop_removal(pressure=pressure, delta=0.0)
+    assert flags[3]
+    assert not flags[[0, 1, 2, 4]].any()
